+++ samples/mvp-sample/src/test/java/org/vaadin/spring/samples/mvp/ApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.vaadin.spring.test.VaadinAppConfiguration;
+import org.vaadin.spring.test.annotation.VaadinAppConfiguration;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringApplicationConfiguration(classes = Application.class)
