+++ spring-vaadin-security/src/test/java/org/vaadin/spring/security/test/DependecyInjectionTest.java
@@ -10,7 +10,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.vaadin.spring.security.VaadinSecurity;
 import org.vaadin.spring.security.VaadinSecurityContext;
-import org.vaadin.spring.test.VaadinAppConfiguration;
+import org.vaadin.spring.test.annotation.VaadinAppConfiguration;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
