+++ spring-vaadin-simple-mvp/src/test/java/org/vaadin/spring/mvp/autowired/AutowiredPresenterIT.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.vaadin.spring.events.EventBus;
-import org.vaadin.spring.test.VaadinAppConfiguration;
+import org.vaadin.spring.test.annotation.VaadinAppConfiguration;
 
 import static org.junit.Assert.assertEquals;
