+++ spring-vaadin-test/src/main/java/org/vaadin/spring/test/annotation/VaadinAppConfiguration.java
@@ -13,13 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaadin.spring.test;
+package org.vaadin.spring.test.annotation;
 
 import org.springframework.test.context.TestExecutionListeners;
 import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
 import org.springframework.test.context.support.DirtiesContextTestExecutionListener;
 import org.springframework.test.context.transaction.TransactionalTestExecutionListener;
 import org.springframework.test.context.web.WebAppConfiguration;
+import org.vaadin.spring.test.VaadinTestExecutionListener;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
